refactor : [#11 서비스 구조 변경 및 명칭 수정 ]
