+++ FeedService/src/main/java/com/example/feedservice/comment/controller/CommentController.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.comment.controller;
+
+public class CommentController {
+}
