+++ FeedService/src/main/java/com/example/feedservice/comment/dto/CommentDto.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.comment.dto;
+
+public class CommentDto {
+}
