+++ FeedService/src/main/java/com/example/feedservice/comment/repository/CommentRepository.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.comment.repository;
+
+public class CommentRepository {
+}
