+++ FeedService/src/main/java/com/example/feedservice/comment/service/CommentService.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.comment.service;
+
+public class CommentService {
+}
