--- FeedService/src/main/java/com/example/feedservice/entity/BaseEntity.java
@@ -1,33 +1,0 @@
-package com.example.feedservice.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.domain.Persistable;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-@Getter
-public abstract class BaseEntity implements Persistable<String> {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdDate;
-
-    @LastModifiedDate
-    private LocalDateTime updatedDate;
-
-    @Override
-    public abstract String getId();
-
-    @Override
-    public boolean isNew() {
-        return createdDate == null;
-    }
-}
