+++ FeedService/src/main/java/com/example/feedservice/feed/controller/FeedController.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.feed.controller;
+
+public class FeedController {
+}
