+++ FeedService/src/main/java/com/example/feedservice/feed/dto/response/RequestDto.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.feed.dto.response;
+
+public class RequestDto {
+}
