+++ FeedService/src/main/java/com/example/feedservice/media/controller/MediaController.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.media.controller;
+
+public class MediaController {
+}
