+++ FeedService/src/main/java/com/example/feedservice/media/dto/MediaDto.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.media.dto;
+
+public class MediaDto {
+}
