+++ FeedService/src/main/java/com/example/feedservice/notification/controller/NotificationController.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.notification.controller;
+
+public class NotificationController {
+}
