+++ FeedService/src/main/java/com/example/feedservice/notification/dto/dto.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.notification.dto;
+
+public class dto {
+}
