+++ FeedService/src/main/java/com/example/feedservice/notification/repository/NotificationRepository.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.notification.repository;
+
+public class NotificationRepository {
+}
