+++ FeedService/src/main/java/com/example/feedservice/notification/service/NotificationService.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.notification.service;
+
+public class NotificationService {
+}
