+++ FeedService/src/main/java/com/example/feedservice/reaction/controller/ReactionController.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.reaction.controller;
+
+public class ReactionController {
+}
