+++ FeedService/src/main/java/com/example/feedservice/reaction/dto/ReactionDto.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.reaction.dto;
+
+public class ReactionDto {
+}
