+++ FeedService/src/main/java/com/example/feedservice/reaction/repository/ReactionRepository.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.reaction.repository;
+
+public class ReactionRepository {
+}
