+++ FeedService/src/main/java/com/example/feedservice/reaction/service/ReactionService.java
@@ -1,0 +1,4 @@
+package com.example.feedservice.reaction.service;
+
+public class ReactionService {
+}
