--- FeedService/src/main/java/com/example/feedservice/repository/PostRepository.java
@@ -1,9 +1,0 @@
-package com.example.feedservice.repository;
-
-import com.example.feedservice.entity.PostEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<PostEntity, String> {
-}
