--- FeedService/src/main/java/com/example/feedservice/util/FeedUtil.java
@@ -1,15 +1,0 @@
-package com.example.feedservice.util;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Component;
-
-import java.util.UUID;
-
-@Component
-public class FeedUtil {
-
-    @Bean
-    public String getUUID() {
-        return UUID.randomUUID().toString();
-    }
-}
